move cache entry class into separate file
