+++ lib/pangea/common/utils/repo_cache.dart
@@ -1,29 +1,5 @@
 import 'package:fluffychat/pangea/common/utils/base_response.dart';
-
-class RepoCacheItem<TResponse extends BaseResponse> {
-  final DateTime timestamp;
-  final TResponse response;
-
-  const RepoCacheItem({required this.timestamp, required this.response});
-
-  bool isExpired(Duration cacheDuration) =>
-      timestamp.isBefore(DateTime.now().subtract(cacheDuration));
-
-  Map<String, dynamic> toJson() => {
-    "timestamp": timestamp.millisecondsSinceEpoch,
-    "response": response.toJson(),
-  };
-
-  factory RepoCacheItem.fromJson(
-    Map<String, dynamic> json, {
-    required TResponse Function(Map<String, dynamic>) responseFromJson,
-  }) {
-    return RepoCacheItem(
-      timestamp: DateTime.fromMillisecondsSinceEpoch(json["timestamp"]),
-      response: responseFromJson(json["response"]),
-    );
-  }
-}
+import 'package:fluffychat/pangea/common/utils/repo_cache_item.dart';
 
 abstract class RepoCache<TResponse extends BaseResponse> {
   Future<void> init();
